+++ app/src/main/java/com/patloew/oeffitracker/data/model/PriceDeduction.kt
@@ -1,6 +1,6 @@
 package com.patloew.oeffitracker.data.model
 
 data class PriceDeduction(
-    val price: Int,
-    val deduction: Int?
+    val price: Long,
+    val deduction: Long?
 )
